refactor booking availability into shared engine
