+++ index.js
@@ -39,6 +39,8 @@ const technicianReworkHelpers = require("./server/technicianRework");
 const technicianReworkIncome = require("./server/services/technicianReworkIncome");
 const adminJobItemsHelpers = require("./server/adminJobItems");
 const customerAvailability = require("./server/services/public/customerAvailability");
+const { registerPublicCustomerAvailabilityRoutes } = require("./server/routes/public/customerAvailability");
+const { registerAdminAvailabilityRoutes } = require("./server/routes/admin/adminAvailability");
 const { createTechnicianJobMoneyHelpers } = require("./server/technicianJobMoneySummary");
 const createSystemRoutes = require("./server/routes/system");
 const createTechnicianDirectoryRoutes = require("./server/routes/users/technicians");
@@ -24615,17 +24617,16 @@ function publicCustomerAvailabilityDeps(db = pool) {
     isTechOffOnDate,
     buildTechWindowsMin,
     listBusyBlocksForTechOnDate,
+    buildBusyIntervalsConservative,
+    buildFreeIntervalsForWindow,
     buildStartIntervalsByCollision,
+    isTechFree,
     toMin,
     minToHHMM,
+    fmtHHMMFromMin,
     getNowBangkokParts,
+    travelBufferMin: TRAVEL_BUFFER_MIN,
   };
-}
-
-function setPublicAvailabilityNoStore(res) {
-  res.set("Cache-Control", "no-store, no-cache, must-revalidate, max-age=0");
-  res.set("Pragma", "no-cache");
-  res.set("Expires", "0");
 }
 
 app.post("/public/pricing_preview", async (req, res) => {
@@ -24674,723 +24675,18 @@ app.post("/public/pricing_preview", async (req, res) => {
 });
 
 
-// =======================================
-// 📅 Availability v2 (รายช่าง + แยก company/partner + ใช้ buffer)
-// =======================================
-app.get("/public/availability_v2", async (req, res) => {
-  setPublicAvailabilityNoStore(res);
-  if (!ENABLE_AVAILABILITY_V2) return res.status(404).json({ error: "DISABLED" });
-
-  const date = (req.query.date || new Date().toISOString().slice(0, 10)).toString();
-  const tech_type = (req.query.tech_type || "company").toString().trim().toLowerCase();
-  // forced=1 (Admin lock): allow showing technicians even if accept_status='paused'
-  const forced = String(req.query.forced || '').trim() === '1';
-  const duration_min = Math.max(15, Number(req.query.duration_min || 60));
-  // crew_size: if a job can be shared by multiple technicians simultaneously,
-  // callers (Admin v2) can request availability based on per-tech workload time.
-  // This is backward compatible: if omitted or invalid, crew_size=1.
-  const crew_size_raw = Number(req.query.crew_size || req.query.crewSize || 1);
-  let crew_size = Math.max(1, Math.min(10, Number.isFinite(crew_size_raw) ? Math.floor(crew_size_raw) : 1));
-  // auto_crew (deprecated): ignored by spec (ห้ามหารเวลา/auto crew)
-  const auto_crew = false;
-  // include_full=1: debug/admin usage to return even unavailable time steps.
-  const include_full = String(req.query.include_full || '').trim() === '1';
-  const slot_step_min = 30;
-  // mode:
-  // - 'start' (default): return blocks of *startable ranges* (เริ่มได้) for the given duration
-  // - 'free': return blocks of *free time* (เวลาว่างจริง) within UI window
-  const mode = String(req.query.mode || req.query.view || 'start').trim().toLowerCase();
-
-  // debug=1 (admin/dev): include backend busy/free + reasons in response
-  const debugFlag = String(req.query.debug || '').trim() === '1';
-  const debugBusy = {};
-  const debugFree = {};
-  const debugReasons = [];
-  const debugInfo = { busy_by_tech: debugBusy, free_by_tech: debugFree, reasons: debugReasons };
-
-  if (!forced) {
-    try {
-      const data = await customerAvailability.computePublicCustomerSlots(
-        publicCustomerAvailabilityDeps(),
-        { ...req.query, date, tech_type, duration_min }
-      );
-      return res.json(data);
-    } catch (e) {
-      const status = Number(e.status || 503);
-      const message = status === 400
-        ? "ข้อมูลบริการไม่ครบสำหรับตรวจคิว กรุณาเลือกรายการบริการใหม่"
-        : "โหลดตารางว่างไม่สำเร็จ";
-      return res.status(status >= 400 && status < 600 ? status : 503).json({ error: message });
-    }
-  }
-
-  // ===== Service Matrix (Option B) =====
-  // criteria from customer booking page (Thai values)
-  const q_job_type = String(req.query.job_type || req.query.jobType || '').trim();
-  const q_ac_type = String(req.query.ac_type || req.query.acType || '').trim();
-  const q_wash_variant = String(req.query.wash_variant || req.query.washVariant || '').trim();
-  const q_repair_variant = String(req.query.repair_variant || req.query.repairVariant || '').trim();
-
-  const normalizeJobKey = (s) => {
-    const v = String(s || '').toLowerCase();
-    if (!v) return null;
-    if (v.includes('ติดตั้ง')) return 'install';
-    if (v.includes('ซ่อม')) return 'repair';
-    if (v.includes('ล้าง')) return 'wash';
-    return null;
-  };
-  const normalizeAcKey = (s) => {
-    const v = String(s || '').toLowerCase();
-    if (!v) return null;
-    if (v.includes('ผนัง') || v.includes('wall')) return 'wall';
-    if (v.includes('สี่ทิศ') || v.includes('4') || v.includes('four')) return 'fourway';
-    if (v.includes('แขวน')) return 'hanging';
-    if (v.includes('ใต้ฝ้า') || v.includes('เปลือย') || v.includes('ฝัง')) return 'ceiling';
-    return null;
-  };
-  const normalizeWashKey = (s) => {
-    const v = String(s || '').toLowerCase();
-    if (!v) return null;
-    if (v.includes('ธรรมดา') || v.includes('normal')) return 'normal';
-    if (v.includes('พรีเมียม') || v.includes('premium')) return 'premium';
-    if (v.includes('แขวนคอย') || v.includes('coil')) return 'coil';
-    if (v.includes('ตัดล้าง') || v.includes('overhaul') || v.includes('ใหญ่')) return 'overhaul';
-    return null;
-  };
-
-  const normalizeRepairKey = (s) => {
-    const v = String(s || '').toLowerCase();
-    if (!v) return null;
-    if (v.includes('à¸£à¸±à¹ˆà¸§') || v.includes('leak')) return 'leak_check';
-    if (v.includes('à¸­à¸°à¹„à¸«à¸¥à¹ˆ') || v.includes('part')) return 'parts';
-    if (v.includes('à¸•à¸£à¸§à¸ˆ') || v.includes('inspect')) return 'inspection';
-    if (v.includes('à¸—à¸±à¹ˆà¸§à¹„à¸›') || v.includes('general')) return 'general';
-    return v.replace(/[^a-z0-9_]+/g, '_').replace(/^_+|_+$/g, '') || null;
-  };
-
-  const criteria = {
-    job: normalizeJobKey(q_job_type),
-    ac: normalizeAcKey(q_ac_type),
-    wash: normalizeWashKey(q_wash_variant),
-    repair: normalizeRepairKey(q_repair_variant),
-    repair_variant: q_repair_variant || null,
-  };
-
-  // Multi-service criteria (customer booking can include multiple AC types in one job)
-  // Query param: services=<JSON.stringify([{job_type,ac_type,wash_variant}...])>
-  let criteriaList = null;
-  try {
-    const sRaw = req.query.services;
-    if (sRaw) {
-      const parsed = JSON.parse(String(sRaw));
-      if (Array.isArray(parsed) && parsed.length) {
-        criteriaList = parsed
-          .map((s) => ({
-            job: normalizeJobKey(s.job_type || q_job_type),
-            ac: normalizeAcKey(s.ac_type),
-            wash: normalizeWashKey(s.wash_variant),
-            repair: normalizeRepairKey(s.repair_variant),
-            repair_variant: String(s.repair_variant || '').trim() || null,
-          }))
-          .filter((c) => c.job && c.ac);
-        if (!criteriaList.length) criteriaList = null;
-      }
-    }
-  } catch (e) {
-    criteriaList = null;
-  }
-
-  const hasCompleteCriteria = (c) => Boolean(
-    c && c.job && c.ac &&
-    !(c.job === 'wash' && c.ac === 'wall' && !c.wash) &&
-    !(c.job === 'repair' && !c.repair_variant)
-  );
-  const hasCriteria = Array.isArray(criteriaList) && criteriaList.length
-    ? criteriaList.every(hasCompleteCriteria)
-    : hasCompleteCriteria(criteria);
-
-  async function loadServiceMatrixMap(usernames) {
-    // NOTE: For customer booking we want strict filtering (no record => not eligible).
-    // But if DB query fails, we fail-open by skipping matrix filtering entirely.
-    loadServiceMatrixMap._ok = true;
-    try {
-      if (!usernames || !usernames.length) return new Map();
-      const r = await pool.query(
-        `SELECT username, matrix_json FROM public.technician_service_matrix WHERE username = ANY($1::text[])`,
-        [usernames]
-      );
-      const m = new Map();
-      for (const row of (r.rows || [])) m.set(String(row.username), row.matrix_json || {});
-      return m;
-    } catch (e) {
-      // fail-open
-      loadServiceMatrixMap._ok = false;
-      console.warn('[availability_v2] loadServiceMatrixMap failed:', e.message);
-      return new Map();
-    }
-  }
-
-  function techMatchesMatrixStrict(matrix, c) {
-    // Strict: missing matrix or missing keys => NOT eligible for customer slots
-    if (!matrix || typeof matrix !== 'object') return false;
-    const mustTrue = (obj, key) => {
-      if (!key) return true;
-      if (!obj || typeof obj !== 'object') return false;
-      return Boolean(obj[key]);
-    };
-    const hasTrue = (obj, keys) => {
-      if (!obj || typeof obj !== 'object') return false;
-      return (keys || []).filter(Boolean).some((key) => Boolean(obj[key]));
-    };
-    if (!mustTrue(matrix.job_types, c.job)) return false;
-    if (!mustTrue(matrix.ac_types, c.ac)) return false;
-    if (c.job === 'wash' && c.ac === 'wall') {
-      if (!mustTrue(matrix.wash_wall_variants, c.wash)) return false;
-    }
-    if (c.job === 'repair' && c.repair_variant) {
-      if (!hasTrue(matrix.repair_variants, [c.repair, c.repair_variant])) return false;
-    }
-    return true;
-  }
-
-  function techMatchesAllCriteriaStrict(matrix, list) {
-    if (!Array.isArray(list) || !list.length) return techMatchesMatrixStrict(matrix, criteria);
-    for (const c of list) {
-      if (!techMatchesMatrixStrict(matrix, c)) return false;
-    }
-    return true;
-  }
-
-  try {
-    // Customer booking must NOT depend on technician "open/close accept jobs" status.
-    // The accept_status (paused/ready) is reserved for the "urgent push" flow only.
-    // Therefore, availability should always include paused technicians.
-    // (Admin lock forced=1 remains meaningful for other rules like weekly off-days handling.)
-    const techsAll = await listTechniciansByType(tech_type, { include_paused: true, allow_type_fallback: forced });
-    // workday overrides (block forced lock on off-days)
-    const offMap = await buildOffMapForDate(date, techsAll.map(t => t.username));
-    const techs = techsAll.filter(t => {
-      // If forced lock mode (admin calendar): include paused technicians.
-      // Still respect explicit overrides (technician_workdays_v2) but
-      // ignore weekly_off_days to prevent bad backfill from hiding all techs.
-      if (forced && isTechOffOnDate(t, date, offMap, { ignoreWeekly: true })) return false;
-      return true;
-    });
-
-    // Customer slot policy (strict, fail-closed):
-    // - Admin must explicitly allow the technician through customer_slot_visible=true.
-    // - The technician must have a service matrix record matching every requested service.
-    // - Missing criteria, missing matrix, or matrix read failure returns no customer slots.
-    // - forced=1 is an admin-only calendar view and intentionally bypasses this customer filter.
-    let techsFiltered = techs;
-    if (!forced) {
-      const visibleTechs = techs.filter(t => t && t.customer_slot_visible === true);
-      if (!hasCriteria) {
-        techsFiltered = [];
-        if (debugFlag) debugReasons.push({ code: 'MISSING_SERVICE_CRITERIA', message: 'ต้องระบุประเภทงานก่อนจึงจะแสดงสลอตลูกค้า' });
-      } else {
-        const matrixMap = await loadServiceMatrixMap(visibleTechs.map(t => t.username));
-        const matrixOk = (loadServiceMatrixMap._ok !== false);
-        techsFiltered = matrixOk
-          ? visibleTechs.filter(t => {
-              const u = String(t.username);
-              if (!matrixMap.has(u)) return false;
-              return techMatchesAllCriteriaStrict(matrixMap.get(u) || null, criteriaList || null);
-            })
-          : [];
-        if (debugFlag && !matrixOk) {
-          debugReasons.push({ code: 'SERVICE_MATRIX_UNAVAILABLE', message: 'โหลดสิทธิ์งานของช่างไม่สำเร็จ จึงปิดสลอตลูกค้าไว้ก่อน' });
-        }
-        if (debugFlag && matrixOk && techsFiltered.length === 0 && visibleTechs.length > 0) {
-          debugReasons.push({ code: 'NO_MATCH_MATRIX', message: 'ไม่มีช่างที่แอดมินอนุญาตและรับงานประเภทนี้' });
-        }
-      }
-    }
-    // special slots map (admin can extend availability)
-    const specialMap = new Map();
-    try {
-      const sr = await pool.query(
-        `SELECT technician_username, start_time, end_time
-         FROM public.technician_special_slots_v2
-         WHERE slot_date=$1::date`,
-        [date]
-      );
-      for (const row of sr.rows) {
-        const u = row.technician_username;
-        if (!specialMap.has(u)) specialMap.set(u, []);
-        specialMap.get(u).push({ start: row.start_time, end: row.end_time });
-      }
-    } catch (e) {
-      // fail-open: do not break availability
-      console.warn("[availability_v2] special slots query failed", e.message);
-    }
-    const tech_count = techsFiltered.length;
-
-    if (debugFlag && tech_count === 0) {
-      debugReasons.push({ code: 'NO_TECH', message: 'ไม่มีช่างที่ตรงเงื่อนไข (tech_type/วันหยุด/สิทธิ์งาน) — หมายเหตุ: โหมดเปิด/ปิดรับงาน (accept_status) ไม่ถูกนำมาใช้กับการจองของลูกค้า' });
-    }
-    // ✅ Crew sizing / parallel work preview
-    // ตามสเปก CWF:
-    // - ห้ามเอา crew_size ไปหารเวลาในโหมด auto/single
-    // - อนุญาตเฉพาะ "preview" โหมด team (เพื่อแสดงข้อมูลเท่านั้น)
-    // ดังนั้น availability จะใช้ duration_min จริงเสมอ (ไม่ divide) เว้นแต่ caller ระบุ preview_team=1 และ assign_mode=team
-    const assign_mode_q = String(req.query.assign_mode || req.query.assignMode || '').trim().toLowerCase();
-    const preview_team = String(req.query.preview_team || req.query.previewTeam || '').trim() === '1';
-    const allowPreviewParallel = preview_team && assign_mode_q === 'team';
-
-    if (!allowPreviewParallel) {
-      crew_size = 1;
-    } else {
-      // preview only: clamp crew_size to [1, tech_count]
-      crew_size = Math.max(1, Math.min(tech_count || 1, Number(crew_size || 1) || 1));
-    }
-
-    // ✅ UI primary window defaults to 09:00–18:00.
-    // Admin forced mode can extend it to include special slots outside normal hours.
-    let uiStartMin = toMin('09:00');
-    let uiEndMin = toMin('18:00');
-    const work_start = '09:00';
-    const work_end = '18:00';
-    if (forced && specialMap.size) {
-      let specialStartMin = uiStartMin;
-      let specialEndMin = uiEndMin;
-      for (const arr of specialMap.values()) {
-        for (const w of (arr || [])) {
-          const ws = toMin(w.start);
-          const we = toMin(w.end);
-          if (Number.isFinite(ws) && Number.isFinite(we) && we > ws) {
-            specialStartMin = Math.min(specialStartMin, ws);
-            specialEndMin = Math.max(specialEndMin, we);
-          }
-        }
-      }
-      uiStartMin = specialStartMin;
-      uiEndMin = specialEndMin;
-    }
-
-    // ✅ If checking "today" in Bangkok time, do not show slots in the past.
-    // Only applies to start-mode slot display; free-mode still shows full-day free/busy ranges.
-    try {
-      const nowBkk = getNowBangkokParts();
-      const todayBkk = `${nowBkk.Y}-${nowBkk.M}-${nowBkk.D}`;
-      if (mode === 'start' && date === todayBkk) {
-        const nowMin = nowBkk.hh * 60 + nowBkk.mm;
-        // round up to slot boundary
-        const step = slot_step_min;
-        const rounded = Math.ceil(nowMin / step) * step;
-        uiStartMin = Math.max(uiStartMin, Math.min(rounded, uiEndMin));
-      }
-    } catch (e) {
-      // ignore timeZone failures
-    }
-
-    // ✅ Duration for collision is ALWAYS the real duration_total (no crew division)
-    const effective_duration_min = Math.max(1, Number(duration_min || 0));
-    const default_effective_block_min = effective_duration_min + TRAVEL_BUFFER_MIN;
-
-    // Build per-tech intervals, then sweep to produce "blocks" (non-fixed steps)
-    const events = new Map(); // min -> { add:[], remove:[] }
-    const addEvent = (min, type, techUser) => {
-      const k = Math.round(min);
-      if (!events.has(k)) events.set(k, { add: [], remove: [] });
-      events.get(k)[type].push(techUser);
-    };
-
-    for (const tech of techsFiltered) {
-      const techWindows = buildTechWindowsMin(tech, date, specialMap, uiStartMin, uiEndMin);
-      if (!techWindows.length) continue;
-
-      const busyBlocks = await listBusyBlocksForTechOnDate(tech.username, date, null);
-
-      // DEBUG: raw job blocks (no buffer) if needed
-      const jobBlocks = null;
-
-      // For each availability window of the technician, compute intervals.
-      for (const w of techWindows) {
-        if (mode === 'free') {
-          // Free blocks: window - busy(with conditional buffer between jobs)
-          const busy = buildBusyIntervalsConservative(busyBlocks);
-          const freeIntervals = buildFreeIntervalsForWindow(busy, w.startMin, w.endMin);
-          if (debugFlag) {
-            debugBusy[tech.username] = (debugBusy[tech.username] || []).concat(busy.map(b=>({ start: fmtHHMMFromMin(b.startMin), end: fmtHHMMFromMin(b.endMin) })));
-            debugFree[tech.username] = (debugFree[tech.username] || []).concat(freeIntervals.map(f=>({ start: fmtHHMMFromMin(f.startMin), end: fmtHHMMFromMin(f.endMin) })));
-          }
-          for (const it of freeIntervals) {
-            // half-open [start, end) -> remove at end
-            addEvent(it.startMin, 'add', tech.username);
-            addEvent(it.endMin, 'remove', tech.username);
-          }
-        } else {
-          // Start ranges: minutes where a job can START (respecting buffer rules)
-          const startIntervals = buildStartIntervalsByCollision(busyBlocks, w.startMin, w.endMin, effective_duration_min);
-        if (debugFlag) {
-          const busy = buildBusyIntervalsConservative(busyBlocks);
-          const free = buildFreeIntervalsForWindow(busy, w.startMin, w.endMin);
-          debugBusy[tech.username] = (debugBusy[tech.username] || []).concat(busy.map(b=>({ start: fmtHHMMFromMin(b.startMin), end: fmtHHMMFromMin(b.endMin) })));
-          debugFree[tech.username] = (debugFree[tech.username] || []).concat(free.map(f=>({ start: fmtHHMMFromMin(f.startMin), end: fmtHHMMFromMin(f.endMin) })));
-        }
-          for (const it of startIntervals) {
-            // Represent as half-open [start, end+1) so we can sweep cleanly in minutes.
-            addEvent(it.startMin, 'add', tech.username);
-            addEvent(it.endMin + 1, 'remove', tech.username);
-          }
-        }
-      }
-    }
-
-    const points = Array.from(events.keys()).sort((a,b)=>a-b);
-    const active = new Set();
-    const slots = [];
-
-    // Ensure deterministic sweep start from uiStartMin
-    const sweepPoints = points.length ? points : [];
-    if (!sweepPoints.length) {
-      if (debugFlag) debugReasons.push({ code: 'NO_EVENTS', message: 'ไม่มีช่วงเวลาว่าง/ช่วงเริ่มงานในหน้าต่างเวลาที่แสดง (อาจเกิดจากวันหยุด/ไม่มี special slot/หรือถูก busy block ทั้งหมด)' });
-      console.log("[availability_v2]", { date, tech_type, forced, duration_min, crew_size, effective_duration_min, tech_count, slots: 0, reason: debugReasons.map(r=>r.code).join(',') });
-      // Public customer response should not reveal technician counts.
-      const isPublicCustomer = !forced;
-      if (isPublicCustomer) {
-        return res.json({
-          date,
-          duration_min: effective_duration_min,
-          slot_step_min,
-          slots: [],
-        });
-      }
-      return res.json({
-        date,
-        tech_type,
-        forced,
-        work_start,
-        work_end,
-        travel_buffer_min: TRAVEL_BUFFER_MIN,
-        duration_min: effective_duration_min,
-        effective_block_min: default_effective_block_min,
-        slot_step_min,
-        tech_count: isPublicCustomer ? undefined : tech_count,
-        crew_size: isPublicCustomer ? undefined : crew_size,
-        slots: [],
-        debug: debugFlag ? debugInfo : undefined,
-      });
-    }
-
-    // Add guard points so sweep covers the whole UI range even if first event starts after uiStartMin
-    if (!events.has(uiStartMin)) events.set(uiStartMin, { add: [], remove: [] });
-    if (!events.has(uiEndMin + 1)) events.set(uiEndMin + 1, { add: [], remove: [] });
-    const pts = Array.from(events.keys()).sort((a,b)=>a-b);
-
-    for (let i=0;i<pts.length;i++) {
-      const t = pts[i];
-      const bucket = events.get(t) || { add: [], remove: [] };
-      // Apply removes first (defensive)
-      for (const u of bucket.remove) active.delete(u);
-      for (const u of bucket.add) active.add(u);
-
-      const next = pts[i+1];
-      if (next == null) continue;
-      const segStart = Math.max(uiStartMin, t);
-      // All internal segments are half-open [t, next).
-      // - start mode: we store end+1 so output should be inclusive (end = next-1)
-      // - free mode: output should stay half-open end (end = next)
-      const segEndExclusive = Math.min(uiEndMin + 1, next);
-      const segEndOut = (mode === 'free') ? Math.min(uiEndMin, segEndExclusive) : (segEndExclusive - 1);
-      if (segEndOut < segStart) continue;
-
-      const ids = Array.from(active);
-      const ok = ids.length >= crew_size;
-      if (!ok && !include_full) continue;
-
-      if (mode === 'start') {
-        // explode "start_range" into fixed step slots so UI shows 09:00-09:30, 09:30-10:00, ...
-        const lastStart = Math.min(segEndOut, uiEndMin - slot_step_min);
-        for (let s = segStart; s <= lastStart; s += slot_step_min) {
-          const e = Math.min(uiEndMin, s + slot_step_min);
-          slots.push({
-            start: minToHHMM(s),
-            end: minToHHMM(e),
-            available: ok,
-            available_tech_ids: ok ? ids : [],
-            capacity: tech_count,
-            available_count: ids.length,
-            crew_size,
-            slot_kind: 'start_step',
-          });
-        }
-      } else {
-        slots.push({
-          start: minToHHMM(segStart),
-          end: minToHHMM(segEndOut),
-          available: ok,
-          available_tech_ids: ok ? ids : [],
-          capacity: tech_count,
-          available_count: ids.length,
-          crew_size,
-          slot_kind: 'free_block',
-        });
-      }
-    }
-
-    if (debugFlag && slots.length === 0 && tech_count > 0) {
-      debugReasons.push({ code: 'BLOCKED', message: 'พบช่างแต่ไม่มีช่วงที่เริ่มงานได้ (ถูก block จาก busy+buffer หรือ duration ยาวเกินช่วงว่าง)' });
-    }
-
-    console.log("[availability_v2]", { date, tech_type, forced, duration_min, crew_size, effective_duration_min, tech_count, slots: slots.length, reason: (debugReasons.length ? debugReasons.map(r=>r.code).join(',') : undefined) });
-
-    // Public customer response should not reveal technician counts or available tech IDs.
-    const isPublicCustomer = !forced;
-    const outSlots = isPublicCustomer
-      ? (slots || []).map(s => ({ start: s.start, end: s.end, available: !!s.available }))
-      : slots;
-
-    if (isPublicCustomer) {
-      return res.json({
-        date,
-        duration_min: effective_duration_min,
-        slot_step_min,
-        slots: outSlots,
-      });
-    }
-
-    res.json({
-      date,
-      tech_type,
-      forced,
-      work_start,
-      work_end,
-      travel_buffer_min: TRAVEL_BUFFER_MIN,
-      duration_min: effective_duration_min,
-      effective_block_min: default_effective_block_min,
-      slot_step_min,
-      tech_count: isPublicCustomer ? undefined : tech_count,
-      crew_size: isPublicCustomer ? undefined : crew_size,
-      mode: (mode === 'free') ? 'free' : 'start',
-      slots: outSlots,
-      debug: debugFlag ? debugInfo : undefined,
-    });
-  } catch (e) {
-    console.error(e);
-    res.status(500).json({ error: "โหลดตารางว่างไม่สำเร็จ" });
-  }
+// Availability route adapters own request parsing and response serialization;
+// the shared engine owns every availability calculation path.
+registerPublicCustomerAvailabilityRoutes(app, {
+  getDependencies: publicCustomerAvailabilityDeps,
+  isEnabled: () => ENABLE_AVAILABILITY_V2,
+  getBangkokTodayYMD,
 });
-
-app.get("/public/availability_calendar_v2", async (req, res) => {
-  setPublicAvailabilityNoStore(res);
-  if (!ENABLE_AVAILABILITY_V2) return res.status(404).json({ error: "DISABLED" });
-  const month = String(req.query.month || "").slice(0, 7);
-  if (!/^\d{4}-\d{2}$/.test(month)) {
-    return res.status(400).json({ error: "กรุณาระบุเดือนให้ถูกต้อง" });
-  }
-  const duration_min = Math.max(15, Number(req.query.duration_min || 60));
-  const tech_type = String(req.query.tech_type || "company").trim().toLowerCase();
-  try {
-    const data = await customerAvailability.computeCalendarSummary(
-      publicCustomerAvailabilityDeps(),
-      { ...req.query, month, tech_type, duration_min }
-    );
-    res.json(data);
-  } catch (e) {
-    const status = Number(e.status || 503);
-    const message = status === 400
-      ? "ข้อมูลบริการไม่ครบสำหรับตรวจคิว กรุณาเลือกรายการบริการใหม่"
-      : "โหลดปฏิทินคิวไม่สำเร็จ";
-    res.status(status >= 400 && status < 600 ? status : 503).json({ error: message });
-  }
+registerAdminAvailabilityRoutes(app, {
+  getDependencies: publicCustomerAvailabilityDeps,
+  isEnabled: () => ENABLE_AVAILABILITY_V2,
+  requireAdminSession,
 });
-
-// Admin: availability by technician (v2) - used for colored rows per tech
-app.get("/admin/availability_by_tech_v2", async (req, res) => {
-  if (!ENABLE_AVAILABILITY_V2) return res.status(404).json({ error: "DISABLED" });
-  const date = (req.query.date || new Date().toISOString().slice(0, 10)).toString();
-  const tech_type = (req.query.tech_type || "company").toString().trim().toLowerCase();
-  const duration_min = Math.max(15, Number(req.query.duration_min || 60));
-  const slot_step_min = 30;
-  try {
-    const include_paused = String(req.query.forced || req.query.include_paused || "").trim() === "1";
-    const techsAll = await listTechniciansByType(tech_type, { include_paused, allow_type_fallback: true });
-    // Same safety as /public/availability_v2:
-    // - respect explicit off override table
-    // - ignore weekly_off_days when include_paused (admin view) to avoid bad backfill hiding all techs
-    const offMap = await buildOffMapForDate(date, (techsAll || []).map(t => t.username));
-    const techs = (techsAll || []).filter(t => {
-      if (include_paused && isTechOffOnDate(t, date, offMap, { ignoreWeekly: true })) return false;
-      return true;
-    });
-    const specialMap = new Map();
-    try {
-      const sr = await pool.query(
-        `SELECT technician_username, start_time, end_time
-         FROM public.technician_special_slots_v2
-         WHERE slot_date=$1::date`,
-        [date]
-      );
-      for (const row of sr.rows) {
-        const u = row.technician_username;
-        if (!specialMap.has(u)) specialMap.set(u, []);
-        specialMap.get(u).push({ start: row.start_time, end: row.end_time });
-      }
-    } catch (e) {
-      console.warn("[admin_availability_by_tech_v2] special slots query failed", e.message);
-    }
-
-    // ✅ Determine global working window:
-let globalStart = toMin("09:00");
-let globalEnd = toMin("18:00");
-
-for (const tech of techs) {
-  const ts = toMin(tech.work_start || "09:00");
-  const te = toMin(tech.work_end || "18:00");
-  if (Number.isFinite(ts)) globalStart = Math.min(globalStart, ts);
-  if (Number.isFinite(te)) globalEnd = Math.max(globalEnd, te);
-  const wins = specialMap.get(tech.username) || [];
-  for (const w of wins) {
-    globalStart = Math.min(globalStart, toMin(w.start));
-    globalEnd = Math.max(globalEnd, toMin(w.end));
-  }
-}
-
-globalStart = Math.max(0, Math.min(24 * 60, globalStart));
-globalEnd = Math.max(0, Math.min(24 * 60, globalEnd));
-
-const work_start = minToHHMM(globalStart);
-const work_end = minToHHMM(globalEnd);
-const startMin = globalStart;
-const endMin = globalEnd;
-
-const default_effective_block_min = Math.max(15, Number(duration_min || 60)) + TRAVEL_BUFFER_MIN;
-
-const all_slots = [];
-for (let t = startMin; t < endMin; t += slot_step_min) {
-  let base = Math.max(15, Number(duration_min || 60));
-  let block = base + TRAVEL_BUFFER_MIN;
-
-  if (t + block > endMin && t + base <= endMin) {
-    block = base;
-  }
-  if (t + base > endMin && t + (base - TRAVEL_BUFFER_MIN) <= endMin) {
-    base = Math.max(15, base - TRAVEL_BUFFER_MIN);
-    block = base;
-  }
-
-  if (t + block > endMin) continue;
-
-  all_slots.push({ start: minToHHMM(t), end: minToHHMM(t + block), service_min: base, block_min: block });
-}
-
-// build per-tech availability
-
-    const techRows = [];
-    for (const tech of techs) {
-      const ts = toMin(tech.work_start || work_start);
-      const te = toMin(tech.work_end || work_end);
-      const wins = specialMap.get(tech.username) || [];
-      const slots = [];
-      for (const s of all_slots) {
-        const t0 = toMin(s.start);
-        let within = (t0 >= ts && t0 + (s.block_min || default_effective_block_min) <= te);
-        if (!within) {
-          for (const w of wins) {
-            const ws = toMin(w.start);
-            const we = toMin(w.end);
-            if (t0 >= ws && t0 + (s.block_min || default_effective_block_min) <= we) { within = true; break; }
-          }
-        }
-        if (!within) {
-          slots.push({ start: s.start, end: s.end, available: false });
-          continue;
-        }
-        const free = await isTechFree(tech.username, `${date}T${s.start}:00`, (s.service_min || duration_min), null);
-        slots.push({ start: s.start, end: s.end, available: !!free });
-      }
-      techRows.push({ username: tech.username, full_name: tech.full_name || null, slots });
-    }
-
-    console.log("[admin_availability_by_tech_v2]", { date, tech_type, duration_min, tech_count: techs.length, slots: all_slots.length });
-    res.json({ date, tech_type, work_start, work_end, duration_min, effective_block_min: default_effective_block_min, slot_step_min, tech_count: techs.length, all_slots, techs: techRows, // aliases for older admin UI code
-      technicians: techs.map(t => ({ username: t.username, full_name: t.full_name || t.username })),
-      slots_by_tech: Object.fromEntries((techRows||[]).map(tr => [tr.username, tr.slots || []]))
-    });
-  } catch (e) {
-    console.error(e);
-    res.status(500).json({ error: "โหลดตารางว่างไม่สำเร็จ" });
-  }
-});
-
-// Defect 7: Admin-only scheduled-availability eligibility diagnostic.
-// Reports each gate (account, employment type, explicit visibility, service matrix,
-// monthly calendar row, advance flag, time window, capacity, collision, final eligible)
-// for a single technician + date + service criteria. Admin-only; never public.
-app.get("/admin/customer-eligibility-diagnostic", requireAdminSession, async (req, res) => {
-  try {
-    const username = String(req.query.username || "").trim();
-    const date = String(req.query.date || "").slice(0, 10);
-    if (!username || !/^\d{4}-\d{2}-\d{2}$/.test(date)) {
-      return res.status(400).json({ error: "ต้องระบุ username และ date (YYYY-MM-DD)" });
-    }
-    const duration_min = Math.max(15, Number(req.query.duration_min || 60));
-    const result = await customerAvailability.diagnoseTechnicianEligibility(
-      publicCustomerAvailabilityDeps(),
-      { ...req.query, username, date, duration_min }
-    );
-    res.json({ ok: true, ...result });
-  } catch (e) {
-    console.error("GET /admin/customer-eligibility-diagnostic error:", e);
-    const status = Number(e.status || 500);
-    res.status(status >= 400 && status < 600 ? status : 500).json({ error: "วิเคราะห์สิทธิ์การแสดงคิวไม่สำเร็จ" });
-  }
-});
-
-app.get("/public/availability", async (req, res) => {
-  const date = (req.query.date || getBangkokTodayYMD()).toString();
-  const start = (req.query.start || "08:00").toString();
-  const end = (req.query.end || "18:00").toString();
-  const slotMin = Math.max(15, Math.min(120, Number(req.query.slot_min || 30)));
-
-  try {
-    // Customer booking availability must NOT depend on technician open/close status.
-    // accept_status is reserved for urgent push offer flow only.
-    const techR = await pool.query(`
-      SELECT u.username
-      FROM public.users u
-      LEFT JOIN public.technician_profiles p ON p.username=u.username
-      WHERE u.role='technician'
-    `);
-    const technicians = (techR.rows || []).map((r) => String(r.username || "").trim()).filter(Boolean);
-    const techCount = technicians.length;
-
-    const toMin = (hhmm) => {
-      const [h, m] = hhmm.split(":").map((x) => Number(x || 0));
-      return h * 60 + m;
-    };
-    const startMin = toMin(start);
-    const endMin = toMin(end);
-
-    const slots = [];
-    for (let t = startMin; t + slotMin <= endMin; t += slotMin) slots.push(t);
-
-    const result = [];
-    for (const t of slots) {
-      const hh = String(Math.floor(t / 60)).padStart(2, "0");
-      const mm = String(t % 60).padStart(2, "0");
-      const iso = `${date}T${hh}:${mm}:00+07:00`;
-      let freeCount = 0;
-      for (const tech of technicians) {
-        if (await isTechFree(tech, iso, slotMin, null)) freeCount++;
-      }
-      result.push({
-        time: `${hh}:${mm}`,
-        available: techCount === 0 ? false : freeCount > 0,
-        capacity: techCount,
-        busy: Math.max(0, techCount - freeCount),
-      });
-    }
-
-    res.json({ date, start, end, slot_min: slotMin, tech_count: techCount, slots: result });
-  } catch (e) {
-    console.error(e);
-    res.status(500).json({ error: "โหลดตารางว่างไม่สำเร็จ" });
-  }
-});
-
 function deriveCustomerScheduledBookingToken(requestKey) {
   const key = String(requestKey || "").trim();
   if (!key) return null;
